requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
anthropic==0.40.0
httpx==0.28.0
beautifulsoup4==4.12.3
lxml>=4.9
python-dotenv==1.0.1
